+++ lib/Excel/Writer/XLSX.pm
@@ -2389,7 +2389,7 @@ Examples:
 
 The width corresponds to the column width value that is specified in Excel. It is approximately equal to the length of a string in the default font of Calibri 11. To set the width in pixels use the C<set_column_pixels()> method, see below.
 
-Unfortunately, there is no way to specify "AutoFit" for a column in the Excel file format. This feature is only available at runtime from within Excel.
+See also the C<autofit()> method to set the column widths based on the data in the column, approximately.
 
 As usual the C<$format> parameter is optional, for additional information, see L</CELL FORMATTING>. If you wish to set the format without changing the width you can pass C<undef> as the width parameter:
 
@@ -2457,6 +2457,33 @@ The option to hide unused rows is used by Excel as an optimisation so that the u
     $worksheet->set_default_row( undef, 1 );
 
 See the C<hide_row_col.pl> example program.
+
+
+=head2 autofit($max_width)
+
+Autofit the worksheet column widths to the widest data in the column, approximately.
+
+    $worksheet->autofit();
+
+Excel autofits columns at runtime when it has access to all of the required worksheet information as well as the Windows functions for calculating display areas based on fonts and formatting. Excel::Writer::XLSX doesn't have access to these Windows functions so it simulates autofit by calculating string widths based on metrics taken from Excel. This isn't perfect but for most cases it should be sufficient and indistinguishable from the output of Excel. However there are some limitations to be aware of when using this method:
+
+=over 4
+
+=item * It is based on the default Excel font type and size of Calibri 11. It will not give accurate results for other fonts or font sizes.
+
+=item * It doesn't take formatting of numbers or dates account, although this may be addressed in a later version.
+
+=item * Autofit is a relatively expensive operation since it performs a calculation for all the populated cells in a worksheet. See the note on performance below.
+
+=back
+
+For cases that don't match your desired output you can set explicit column widths via C<set_column()> or C<set_column_pixels()> method ignores columns that have already been explicitly set if the width is greater than the calculated autofit width. Alternatively, setting the column width explicitly after calling C<autofit()> will override the autofit value. You can also set an upper limit using the optional C<$max_width> parameter as explained below.
+
+Excel autofits very long strings up to limit of 1790 pixels/255 characters. This is often too wide to display on a single screen at normal zoom. As such the optional C<$max_width> parameter is provided to enable a smaller upper pixel limit for autofitting long strings. A value of 300 pixels is recommended as a good compromise between column width and readability:
+
+    $worksheet->autofit(300);
+
+If you need more control over the autofit effect you can use the L<Excel::Writer::XLSX::Utility> C<xl_cell_autofit_width()> function to calculate the autofit width for a cell value. This is the same calculation used internally by C<autofit()>.